requirements.txt
streamlit>=1.42.0
PyJWT==2.10.1
# Dash app (app_2.py)
dash>=2.16.0
dash-bootstrap-components>=1.6.0
requests-oauthlib>=1.3.0
toml>=0.10.2
